refactor: createProjectMembershipsOnSignup
