+++ src/features/auth/lib/emailPassword.ts
@@ -1,5 +1,4 @@
-import { env } from "@/src/env.mjs";
-import { processMembershipInvitations } from "@/src/features/auth/lib/processMembershipInvitations";
+import { createProjectMembershipsOnSignup } from "@/src/features/auth/lib/createProjectMembershipsOnSignup";
 import { prisma } from "@/src/server/db";
 import { compare, hash } from "bcryptjs";
 
@@ -31,37 +30,15 @@ export async function createUserEmailPassword(
     throw new Error("User with email already exists. Please sign in.");
   }
 
-  // set demoProjectId if env exists and project exists in db
-  const demoProjectId = env.NEXT_PUBLIC_DEMO_PROJECT_ID
-    ? (
-        await prisma.project.findUnique({
-          where: {
-            id: env.NEXT_PUBLIC_DEMO_PROJECT_ID,
-          },
-        })
-      )?.id
-    : undefined;
-
   const newUser = await prisma.user.create({
     data: {
       email: email.toLowerCase(),
       password: hashedPassword,
       name,
-      // if demo project id is set grant user access to it
-      ...(demoProjectId
-        ? {
-            memberships: {
-              create: {
-                projectId: demoProjectId,
-                role: "VIEWER",
-              },
-            },
-          }
-        : undefined),
     },
   });
 
-  await processMembershipInvitations(email, newUser.id);
+  await createProjectMembershipsOnSignup(newUser);
 
   return newUser.id;
 }
