+++ src/server/auth.ts
@@ -8,16 +8,16 @@ import {
 import { PrismaAdapter } from "@next-auth/prisma-adapter";
 import { prisma } from "@/src/server/db";
 import { verifyPassword } from "@/src/features/auth/lib/emailPassword";
-import { processMembershipInvitations } from "@/src/features/auth/lib/processMembershipInvitations";
 import { parseFlags } from "@/src/features/feature-flags/utils";
 import { env } from "@/src/env.mjs";
+import { createProjectMembershipsOnSignup } from "@/src/features/auth/lib/createProjectMembershipsOnSignup";
+import { type Adapter } from "next-auth/adapters";
 
 // Providers
 import { type Provider } from "next-auth/providers";
 import CredentialsProvider from "next-auth/providers/credentials";
 import GoogleProvider from "next-auth/providers/google";
 import GitHubProvider from "next-auth/providers/github";
-import { type Adapter } from "next-auth/adapters";
 
 // Use secure cookies on https hostnames, exception for Vercel which sets NEXTAUTH_URL without the protocol
 const useSecureCookies =
@@ -128,47 +128,7 @@ const extendedPrismaAdapter: Adapter = {
 
     const user = await prismaAdapter.createUser(profile);
 
-    // Demo project access
-    const demoProjectId = env.NEXT_PUBLIC_DEMO_PROJECT_ID
-      ? (
-          await prisma.project.findUnique({
-            where: {
-              id: env.NEXT_PUBLIC_DEMO_PROJECT_ID,
-            },
-          })
-        )?.id
-      : undefined;
-    if (demoProjectId !== undefined) {
-      await prisma.membership.create({
-        data: {
-          projectId: demoProjectId,
-          userId: user.id,
-          role: "VIEWER",
-        },
-      });
-    }
-
-    await processMembershipInvitations(user.email, user.id);
-
-    // set default project access
-    const defaultProjectID = env.LANGFUSE_DEFAULT_PROJECT_ID
-      ? (
-          await prisma.project.findUnique({
-            where: {
-              id: env.LANGFUSE_DEFAULT_PROJECT_ID,
-            },
-          })
-        )?.id
-      : undefined;
-    if (defaultProjectID !== undefined) {
-      await prisma.membership.create({
-        data: {
-          projectId: defaultProjectID,
-          userId: user.id,
-          role: env.LANGFUSE_DEFAULT_PROJECT_ROLE ?? "VIEWER",
-        },
-      });
-    }
+    await createProjectMembershipsOnSignup(user);
 
     return user;
   },
